requirements.txt
asgiref==3.2.3
astroid==2.5.1
Click==7.0
colorama==0.4.4
Command==0.1.0
Django==3.0.3
djangorestframework==3.12.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.0
pandas==0.25.3
path==13.1.0
pycryptodome==3.9.7
pylint==2.7.2
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
sqlparse==0.3.0
toml==0.10.2
Werkzeug==0.16.0
wrapt==1.12.1
